Add root mode
For convenience, the crawler now grows from a single "root edge,"
which doesn't represent an edge of the Cayley graph. It binds together
the climbers sprouting from the central tile. In the future, it will
provide access to things like drawing methods and a list of all the
home maps in the tree.