Improve orbit pruning
Before, `orbiter` returned an empty context for each triangle too small
to be worth drawing, so the list of triangle fills had a context for
each triangle in the orbit, even if it wasn't drawn. This used too much
memory when drawing a caterpillar lamination with ascent = 3.

Now, `orbiter` returns nothing for each triangle too small to draw, and
mapcollect has the ability to prune nothing values from the final list.
This makes makes a caterpillar lamination with ascent = 3 drawable.