Customize Compose to get paths working
Compose 0.7.0, the release that added support for Julia 1, also
removed the `path` primitive. I set up a Pkg environment with a
customized Compose package and tweaked version 0.7.4 to get `path`
back. My tweaks are recorded as commit 51e5960 in the local Compose
repository.

Until now, for the record, I was running Compose 0.5.5 in Julia 0.5.2.