Hack together lamination-drawing for real projective structures
Add a transition operator type parameter to Cocycle and associated
types. The parameter is just an integer for now, but that should be
refined. Dispatch on the type parameter to draw appropriate lamination
pictures. (There are a lot of holes in the dispatch system for now.)

At double precision (Float64), the stable lines of SL(3,R) elements can
go completely haywire even just a few steps out on the Cayley graph.
I've temporarily turned up the precision of all matrix computations to
get usable results. This will hurt performance in the SL(2,C) case, so
the precision needs to be made adjustable in the end.

Abelianization isn't implemented yet.